backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.1
langchain==0.1.17
langchain-community==0.0.37
langchain-text-splitters==0.0.1
pinecone-client==2.2.4
sentence-transformers==2.6.1
llama-cpp-python==0.2.75
python-dotenv==1.0.1

